+++ src/refactoring/features/admin/components/ProductManager.tsx
@@ -1,50 +1,127 @@
-import { Coupon, Discount, Product } from "../../../../types";
-import { formatDiscountRate } from "../../../utils";
+import { useState } from "react";
+import { Discount, Product } from "../../../../types";
+import { formatDiscountRate, validateProductData } from "../../../utils";
 
 interface Props {
   products: Product[];
   onProductUpdate: (updatedProduct: Product) => void;
   onProductAdd: (newProduct: Product) => void;
-  onCouponAdd: (newCoupon: Coupon) => void;
-  setShowNewProductForm: (show: boolean) => void;
-  showNewProductForm: boolean;
-  newProduct: Omit<Product, "id">;
-  setNewProduct: (newProduct: Omit<Product, "id">) => void;
-  newDiscount: Discount;
-  setNewDiscount: (newDiscount: Discount) => void;
-  openProductIds: Set<string>;
-  handleEditProduct: (product: Product) => void;
-  handleRemoveDiscount: (productId: string, index: number) => void;
-  handleAddDiscount: (productId: string) => void;
-  handleAddNewProduct: () => void;
-  handleProductNameUpdate: (productId: string, newName: string) => void;
-  handlePriceUpdate: (productId: string, newPrice: number) => void;
-  handleStockUpdate: (productId: string, newStock: number) => void;
-  toggleProductAccordion: (productId: string) => void;
-  editingProduct: Product | null;
-  handleEditComplete: () => void;
 }
 
 export const ProductManager = ({
   products,
-  setShowNewProductForm,
-  showNewProductForm,
-  newProduct,
-  setNewProduct,
-  newDiscount,
-  setNewDiscount,
-  openProductIds,
-  handleEditProduct,
-  handleRemoveDiscount,
-  handleAddDiscount,
-  handleAddNewProduct,
-  handleProductNameUpdate,
-  handlePriceUpdate,
-  handleStockUpdate,
-  toggleProductAccordion,
-  editingProduct,
-  handleEditComplete
+  onProductUpdate,
+  onProductAdd
 }: Props) => {
+  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
+  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
+  const [newDiscount, setNewDiscount] = useState<Discount>({
+    quantity: 0,
+    rate: 0
+  });
+  const [showNewProductForm, setShowNewProductForm] = useState(false);
+  const [newProduct, setNewProduct] = useState<Omit<Product, "id">>({
+    name: "",
+    price: 0,
+    stock: 0,
+    discounts: []
+  });
+
+  const toggleProductAccordion = (productId: string) => {
+    setOpenProductIds((prev) => {
+      const newSet = new Set(prev);
+      if (newSet.has(productId)) {
+        newSet.delete(productId);
+      } else {
+        newSet.add(productId);
+      }
+      return newSet;
+    });
+  };
+
+  // handleEditProduct 함수 수정
+  const handleEditProduct = (product: Product) => {
+    setEditingProduct({ ...product });
+  };
+
+  // 새로운 핸들러 함수 추가
+  const handleProductNameUpdate = (productId: string, newName: string) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const updatedProduct = { ...editingProduct, name: newName };
+      setEditingProduct(updatedProduct);
+    }
+  };
+
+  // 새로운 핸들러 함수 추가
+  const handlePriceUpdate = (productId: string, newPrice: number) => {
+    if (editingProduct && editingProduct.id === productId) {
+      const updatedProduct = { ...editingProduct, price: newPrice };
+      setEditingProduct(updatedProduct);
+    }
+  };
+
+  // 수정 완료 핸들러 함수 추가
+  const handleEditComplete = () => {
+    if (editingProduct) {
+      onProductUpdate(editingProduct);
+      setEditingProduct(null);
+    }
+  };
+
+  const handleStockUpdate = (productId: string, newStock: number) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct) {
+      const newProduct = { ...updatedProduct, stock: newStock };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+    }
+  };
+
+  const handleAddDiscount = (productId: string) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct && editingProduct) {
+      const newProduct = {
+        ...updatedProduct,
+        discounts: [...updatedProduct.discounts, newDiscount]
+      };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+      setNewDiscount({ quantity: 0, rate: 0 });
+    }
+  };
+
+  const handleRemoveDiscount = (productId: string, index: number) => {
+    const updatedProduct = products.find((p) => p.id === productId);
+    if (updatedProduct) {
+      const newProduct = {
+        ...updatedProduct,
+        discounts: updatedProduct.discounts.filter((_, i) => i !== index)
+      };
+      onProductUpdate(newProduct);
+      setEditingProduct(newProduct);
+    }
+  };
+
+  const handleAddNewProduct = () => {
+    const productWithId = { ...newProduct, id: Date.now().toString() };
+
+    const validationErrors = validateProductData(productWithId);
+
+    if (validationErrors.length > 0) {
+      alert(validationErrors.map((error) => error.message).join("\n"));
+      return;
+    }
+
+    onProductAdd(productWithId);
+    setNewProduct({
+      name: "",
+      price: 0,
+      stock: 0,
+      discounts: []
+    });
+    setShowNewProductForm(false);
+  };
+
   return (
     <div>
       <h2 className="text-2xl font-semibold mb-4">상품 관리</h2>
