+++ src/refactoring/pages/admin/AdminPage.tsx
@@ -1,7 +1,5 @@
-import { useState } from "react";
-import { Coupon, Discount, Product } from "../../../types";
+import { Coupon, Product } from "../../../types";
 import { PageLayout } from "../../components";
-import { validateProductData } from "../../utils";
 import { CouponManager, ProductManager } from "../../features/admin/components";
 
 interface Props {
@@ -19,162 +17,14 @@ export const AdminPage = ({
   onProductAdd,
   onCouponAdd
 }: Props) => {
-  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
-  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
-  const [newDiscount, setNewDiscount] = useState<Discount>({
-    quantity: 0,
-    rate: 0
-  });
-  const [newCoupon, setNewCoupon] = useState<Coupon>({
-    name: "",
-    code: "",
-    discountType: "percentage",
-    discountValue: 0
-  });
-  const [showNewProductForm, setShowNewProductForm] = useState(false);
-  const [newProduct, setNewProduct] = useState<Omit<Product, "id">>({
-    name: "",
-    price: 0,
-    stock: 0,
-    discounts: []
-  });
-
-  const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(productId)) {
-        newSet.delete(productId);
-      } else {
-        newSet.add(productId);
-      }
-      return newSet;
-    });
-  };
-
-  // handleEditProduct 함수 수정
-  const handleEditProduct = (product: Product) => {
-    setEditingProduct({ ...product });
-  };
-
-  // 새로운 핸들러 함수 추가
-  const handleProductNameUpdate = (productId: string, newName: string) => {
-    if (editingProduct && editingProduct.id === productId) {
-      const updatedProduct = { ...editingProduct, name: newName };
-      setEditingProduct(updatedProduct);
-    }
-  };
-
-  // 새로운 핸들러 함수 추가
-  const handlePriceUpdate = (productId: string, newPrice: number) => {
-    if (editingProduct && editingProduct.id === productId) {
-      const updatedProduct = { ...editingProduct, price: newPrice };
-      setEditingProduct(updatedProduct);
-    }
-  };
-
-  // 수정 완료 핸들러 함수 추가
-  const handleEditComplete = () => {
-    if (editingProduct) {
-      onProductUpdate(editingProduct);
-      setEditingProduct(null);
-    }
-  };
-
-  const handleStockUpdate = (productId: string, newStock: number) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct) {
-      const newProduct = { ...updatedProduct, stock: newStock };
-      onProductUpdate(newProduct);
-      setEditingProduct(newProduct);
-    }
-  };
-
-  const handleAddDiscount = (productId: string) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct && editingProduct) {
-      const newProduct = {
-        ...updatedProduct,
-        discounts: [...updatedProduct.discounts, newDiscount]
-      };
-      onProductUpdate(newProduct);
-      setEditingProduct(newProduct);
-      setNewDiscount({ quantity: 0, rate: 0 });
-    }
-  };
-
-  const handleRemoveDiscount = (productId: string, index: number) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct) {
-      const newProduct = {
-        ...updatedProduct,
-        discounts: updatedProduct.discounts.filter((_, i) => i !== index)
-      };
-      onProductUpdate(newProduct);
-      setEditingProduct(newProduct);
-    }
-  };
-
-  const handleAddCoupon = () => {
-    onCouponAdd(newCoupon);
-    setNewCoupon({
-      name: "",
-      code: "",
-      discountType: "percentage",
-      discountValue: 0
-    });
-  };
-
-  const handleAddNewProduct = () => {
-    const productWithId = { ...newProduct, id: Date.now().toString() };
-
-    const validationErrors = validateProductData(productWithId);
-
-    if (validationErrors.length > 0) {
-      alert(validationErrors.map((error) => error.message).join("\n"));
-      return;
-    }
-
-    onProductAdd(productWithId);
-    setNewProduct({
-      name: "",
-      price: 0,
-      stock: 0,
-      discounts: []
-    });
-    setShowNewProductForm(false);
-  };
-
   return (
     <PageLayout title="상품 관리">
       <ProductManager
         products={products}
         onProductUpdate={onProductUpdate}
         onProductAdd={onProductAdd}
-        onCouponAdd={onCouponAdd}
-        setShowNewProductForm={setShowNewProductForm}
-        showNewProductForm={showNewProductForm}
-        newProduct={newProduct}
-        setNewProduct={setNewProduct}
-        newDiscount={newDiscount}
-        setNewDiscount={setNewDiscount}
-        openProductIds={openProductIds}
-        handleEditProduct={handleEditProduct}
-        handleRemoveDiscount={handleRemoveDiscount}
-        handleAddDiscount={handleAddDiscount}
-        handleAddNewProduct={handleAddNewProduct}
-        handleProductNameUpdate={handleProductNameUpdate}
-        handlePriceUpdate={handlePriceUpdate}
-        handleStockUpdate={handleStockUpdate}
-        toggleProductAccordion={toggleProductAccordion}
-        editingProduct={editingProduct}
-        handleEditComplete={handleEditComplete}
       />
-      <CouponManager
-        newCoupon={newCoupon}
-        setNewCoupon={setNewCoupon}
-        handleAddCoupon={handleAddCoupon}
-        coupons={coupons}
-      />
+      <CouponManager onCouponAdd={onCouponAdd} coupons={coupons} />
     </PageLayout>
   );
 };
